requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
matplotlib==3.8.4
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.4.2
seaborn==0.13.2
SQLAlchemy==2.0.30
